refactor: move `FileInfo` to its own file
