+++ etc/index.api.md
@@ -35,7 +35,6 @@ interface Document_2 {
     alias: string[];
     // (undocumented)
     attributes: NormalizedDocumentAttributes;
-    // (undocumented)
     body: string;
     // (undocumented)
     fileInfo: FileInfo;
