+++ src/document/document.ts
@@ -1,4 +1,5 @@
 import { type DocumentOutline } from "../document-outline";
+import { type FileInfo } from "./file-info";
 
 /**
  * @internal
@@ -90,26 +91,6 @@ export interface NormalizedDocumentAttributes {
 /**
  * @public
  */
-export interface FileInfo {
-    /** path relative to configured base */
-    path: string;
-    /** filename without extension */
-    name: string;
-    /** path relative to project root */
-    fullPath: string;
-    /**
-     * Output filename or `false` to disable writing result to a file.
-     *
-     * When passing a filename these optional placeholders can be used:
-     *
-     * - `[hash]` - replaced with the hash of the content body.
-     */
-    outputName: string | false;
-}
-
-/**
- * @public
- */
 export interface Document {
     /** unique identifier for this document */
     id: string;
@@ -125,6 +106,7 @@ export interface Document {
 
     attributes: NormalizedDocumentAttributes;
 
+    /** content of this document */
     body: string;
 
     /** Document outline (i.e. the heading structure) */
